refactor(wrappers): fold the flat format into ocp_serialize, drop ocp_wire
With the wire format reduced to base64 BREP in plain JSON, a dedicated
serialization module with its own 300-line test suite is no longer
warranted. Move the format's functions (shape<->brep, the shape/assembly
objects, encode/decode, dumps/loads/serialize/deserialize) into
ocp_serialize - which already owns downcast()/shapetype() and is the
natural serialization home - and delete ocp_wire.py and test_ocp_wire.py.

wrapper_common and the 15 main-process callers now use ocp_serialize.
The format is covered by the part, sketch, assembly, convert-matrix and
render suites that exercise the real wrapper pipe (91 passed), so no
standalone format test remains.

No behavior change: same flat BREP objects, same single-line JSON
envelope.
