+++ partcad/src/partcad/cache_shape.py
@@ -17,7 +17,7 @@ from .utils import total_size
 from . import telemetry
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "wrappers"))
-import ocp_wire
+import ocp_serialize
 
 # The on-disk shape cache used to be pickled. That worked only because the
 # wrapper protocol installed OCP 'copyreg' handlers as a global side effect -
@@ -46,7 +46,7 @@ class ShapeCache(Cache):
                 # else:
                 serialization = self.serialization
                 if serialization == SERIALIZATION_JSON:
-                    data = ocp_wire.dumps(value).encode("utf-8")
+                    data = ocp_serialize.dumps(value).encode("utf-8")
                 elif serialization == SERIALIZATION_BREP:
                     data = Shape.to_brep(value)
                 else:
@@ -112,7 +112,7 @@ class ShapeCache(Cache):
             serialization = self.serialization
             if serialization == SERIALIZATION_JSON:
                 try:
-                    obj = ocp_wire.loads(data)
+                    obj = ocp_serialize.loads(data)
                 except:
                     results[key] = None
                     in_memory[key] = False
