+++ partcad/src/partcad/part_factory_brep.py
@@ -12,7 +12,7 @@ from .exception import FileReadError, PartFactoryError
 from . import telemetry
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "wrappers"))
-import ocp_wire
+import ocp_serialize
 
 
 @telemetry.instrument()
@@ -111,7 +111,7 @@ class PartFactoryBrep(PartFactoryFile):
         request = {"build_parameters": {}}
 
         # Serialize the request
-        request_serialized = ocp_wire.serialize(request)
+        request_serialized = ocp_serialize.serialize(request)
 
         # Run the subprocess and handle the response
         try:
@@ -131,7 +131,7 @@ class PartFactoryBrep(PartFactoryFile):
                 pc_logging.error(errors)
                 raise Exception(errors)
 
-            response = ocp_wire.deserialize(response_serialized)
+            response = ocp_serialize.deserialize(response_serialized)
             if not response.get("success", False):
                 pc_logging.error(response["exception"])
                 raise PartFactoryError(response["exception"])
