+++ partcad/src/partcad/part_factory_build123d.py
@@ -25,7 +25,7 @@ from . import wrapper
 from . import logging as pc_logging
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "wrappers"))
-import ocp_wire
+import ocp_serialize
 
 from . import telemetry
 
@@ -83,8 +83,8 @@ class PartFactoryBuild123d(PartFactoryPython):
             request["patch"] = patch
 
             # Serialize the request
-            with telemetry.start_as_current_span("*PartFactoryBuild123d.instantiate.{ocp_wire.serialize}"):
-                request_serialized = ocp_wire.serialize(request)
+            with telemetry.start_as_current_span("*PartFactoryBuild123d.instantiate.{ocp_serialize.serialize}"):
+                request_serialized = ocp_serialize.serialize(request)
 
             # TODO: @alexanderilyin: those should be read from the package/part config?
             await self.runtime.ensure_async(
@@ -136,7 +136,7 @@ class PartFactoryBuild123d(PartFactoryPython):
 
             try:
                 # pc_logging.error("Response: %s" % response_serialized)
-                result = ocp_wire.deserialize(response_serialized)
+                result = ocp_serialize.deserialize(response_serialized)
             except Exception as e:
                 part.error("Exception while deserializing %s: %s" % (part.name, e))
                 return None
