+++ partcad/src/partcad/part_factory_scad.py
@@ -20,7 +20,7 @@ from . import telemetry
 from . import wrapper
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "wrappers"))
-import ocp_wire
+import ocp_serialize
 
 
 @telemetry.instrument()
@@ -106,7 +106,7 @@ class PartFactoryScad(PartFactoryFile):
             wrapper_path = wrapper.get("import_mesh.py")
 
             request = {"fallback_import_stl": True}
-            request_serialized = ocp_wire.serialize(request)
+            request_serialized = ocp_serialize.serialize(request)
 
             await self.runtime.ensure_async("ocp-tessellate==3.0.9")
             await self.runtime.ensure_async("typing_extensions==4.12.2")
@@ -132,7 +132,7 @@ class PartFactoryScad(PartFactoryFile):
                 return None
 
             try:
-                result = ocp_wire.deserialize(response_serialized)
+                result = ocp_serialize.deserialize(response_serialized)
             except Exception as e:
                 part.error("%s: %s" % (part.name, e))
                 return None
