+++ partcad/src/partcad/sketch_factory_cadquery.py
@@ -25,7 +25,7 @@ from . import wrapper
 from . import logging as pc_logging
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "wrappers"))
-import ocp_wire
+import ocp_serialize
 
 from . import telemetry
 
@@ -79,7 +79,7 @@ class SketchFactoryCadquery(SketchFactoryPython):
             request["patch"] = patch
 
             # Serialize the request
-            request_serialized = ocp_wire.serialize(request)
+            request_serialized = ocp_serialize.serialize(request)
 
             await self.runtime.ensure_async(
                 "ocp-tessellate==3.0.9",
@@ -128,7 +128,7 @@ class SketchFactoryCadquery(SketchFactoryPython):
                     sketch.error("%s: %s" % (sketch.name, error_line))
 
             try:
-                result = ocp_wire.deserialize(response_serialized)
+                result = ocp_serialize.deserialize(response_serialized)
                 pc_logging.debug("Response: %s" % result)
             except Exception as e:
                 sketch.error("Exception while deserializing %s: %s" % (sketch.name, e))
