+++ partcad/src/partcad/sketch_factory_dxf.py
@@ -15,7 +15,7 @@ from . import logging as pc_logging
 from .sketch_factory_python import SketchFactoryPython
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "wrappers"))
-import ocp_wire
+import ocp_serialize
 
 from . import telemetry
 
@@ -75,7 +75,7 @@ class SketchFactoryDxf(SketchFactoryPython):
                     "include": self.include,
                     "exclude": self.exclude,
                 }
-                request_serialized = ocp_wire.serialize(request)
+                request_serialized = ocp_serialize.serialize(request)
 
                 await self.runtime.ensure_async("cadquery-ocp==7.7.2")
                 await self.runtime.ensure_async("cadquery==2.5.2")
@@ -95,7 +95,7 @@ class SketchFactoryDxf(SketchFactoryPython):
                     pc_logging.error(errors)
                     raise Exception(errors)
 
-                result = ocp_wire.deserialize(response_serialized)
+                result = ocp_serialize.deserialize(response_serialized)
 
                 if not result["success"]:
                     pc_logging.error(result["exception"])
